hotfix: weather

+++ src/main/java/sevenstar/marineleisure/activity/dto/request/ActivityWeatherRequest.java
@@ -3,7 +3,7 @@ package sevenstar.marineleisure.activity.dto.request;
 import java.math.BigDecimal;
 
 public record ActivityWeatherRequest(
-    BigDecimal latitude,
-    BigDecimal longitude
+    Float latitude,
+	Float longitude
 ) {
 }
